+++ xrpl/utils/__init__.py
@@ -9,6 +9,7 @@ from xrpl.utils.time_conversions import (
     ripple_time_to_datetime,
     ripple_time_to_posix,
 )
+from xrpl.utils.txn_parser import parse_balance_changes
 from xrpl.utils.xrp_conversions import XRPRangeException, drops_to_xrp, xrp_to_drops
 
 __all__ = [
@@ -23,4 +24,5 @@ __all__ = [
     "XRPRangeException",
     "XRPLTimeRangeException",
     "create_cross_chain_payment",
+    "parse_balance_changes",
 ]